Update README.md
DAONOW is a collection of source code for the creation of DAO<br />
DAONOW is a platform dedicated to the development and management of Decentralized Autonomous Organizations (DAOs), emphasizing the "code is law" philosophy. This approach ensures trustless, immutable exchanges by embedding governance and operational rules directly into smart contracts, minimizing the need for traditional trust mechanisms. DAOs play a crucial role in decentralized finance (DeFi) and blockchain ecosystems, enabling collective decision-making and resource management in a transparent, secure manner​​.